mongo calss name rename
