+++ modules/blogging/test/Volo.Blogging.MongoDB.Tests/Volo/Blogging/Blogs/BlogRepository_Tests.cs
@@ -1,8 +1,9 @@
 ﻿using Volo.Blogging.Blogs;
+using Volo.Blogging.MongoDB;
 
 namespace Volo.Blogging
 {
-    public class BlogRepository_Tests : BlogRepository_Tests<BloggingMongoDBTestModule>
+    public class BlogRepository_Tests : BlogRepository_Tests<BloggingMongoDbTestModule>
     {
     }
 }
