+++ modules/blogging/test/Volo.Blogging.MongoDB.Tests/Volo/Blogging/Comments/CommentRepository_Tests.cs
@@ -1,8 +1,9 @@
 ﻿using Volo.Blogging.Comments;
+using Volo.Blogging.MongoDB;
 
 namespace Volo.Blogging
 {
-    public class CommentRepository_Tests : CommentRepository_Tests<BloggingMongoDBTestModule>
+    public class CommentRepository_Tests : CommentRepository_Tests<BloggingMongoDbTestModule>
     {
     }
 }
