+++ modules/blogging/test/Volo.Blogging.MongoDB.Tests/Volo/Blogging/Posts/PostRepository_Tests.cs
@@ -1,8 +1,9 @@
 ﻿using Volo.Blogging.Posts;
+using Volo.Blogging.MongoDB;
 
 namespace Volo.Blogging
 {
-    public class PostRepository_Tests : PostRepository_Tests<BloggingMongoDBTestModule>
+    public class PostRepository_Tests : PostRepository_Tests<BloggingMongoDbTestModule>
     {
     }
 }
