+++ modules/blogging/test/Volo.Blogging.MongoDB.Tests/Volo/Blogging/Tagging/TagRepository_Tests.cs
@@ -1,8 +1,9 @@
 ﻿using Volo.Blogging.Tagging;
+using Volo.Blogging.MongoDB;
 
 namespace Volo.Blogging
 {
-    public class TagRepository_Tests : TagRepository_Tests<BloggingMongoDBTestModule>
+    public class TagRepository_Tests : TagRepository_Tests<BloggingMongoDbTestModule>
     {
     }
 }
